CAMBIOS

+++ src/main/java/mx/uady/appbusqueda/config/JwtTokenUtil.java
@@ -21,7 +21,7 @@ public class JwtTokenUtil implements Serializable {
 
 	private static final long serialVersionUID = -2550185165626007488L;
 
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 50000;
+	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * (long)50000;
 
 	//@Value("${jwt.secret}")
 	//private String secret;
